+++ CHANGELOG.md
@@ -27,6 +27,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
     unordered.
 
 ### Fixed
+- `cache::Cache::required_object`: Fix required object resolution when merges are involved.
 
 
 ## 0.2.2